\name{morphology-package}
\alias{morphology-package}
\alias{morphology}
\docType{package}
\title{
\packageTitle{morphology}
}
\description{
\packageDescription{morphology}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{morphology}
\packageIndices{morphology}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{morphology}

Maintainer: \packageMaintainer{morphology}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
